break out into components
